pyproject.toml
[project]
name = "mcq-generator"
version = "0.1.0"
description = "RAG based MCQ Generator"
authors = [
    {name = "Author",email = "author@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "prisma (>=0.15.0,<0.16.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-google-genai (>=4.2.2,<5.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "chromadb (>=1.5.8,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "pydantic (>=2.13.1,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "cloudinary (>=1.44.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
